docs: move the latest-commit fetch into the app store
AppSettingsCommits and AppFooter carried the same octokit fetch —
the settings copy TTL-cached, the footer copy not, so the footer
re-hit the GitHub API on every intersection. One fetchCommit() action
owns the cache; the footer now inherits it (verified: one commits
request across two page loads).
