+++ apps/docs/src/components/app/settings/AppSettingsCommits.vue
@@ -1,47 +1,15 @@
 <script setup lang="ts">
-  // Framework
-  import { createStorage, useLogger } from '@vuetify/v0'
-
-  import { CACHE_TTL } from '@/constants/cache'
-
   // Stores
   import { useAppStore } from '@/stores/app'
 
   // Utilities
   import { onMounted } from 'vue'
 
-  const storage = createStorage({ prefix: 'v0-commit:', ttl: CACHE_TTL })
-
   const app = useAppStore()
-  const logger = useLogger()
 
   const buildSha = import.meta.env.VITE_GITHUB_SHA as string | undefined
 
-  onMounted(async () => {
-    if (app.stats.commit) return
-
-    const cached = storage.get<typeof app.stats.commit | null>('latest', null)
-    if (cached.value) {
-      app.stats.commit = cached.value
-      return
-    }
-
-    try {
-      const octokit = await import('@/plugins/octokit').then(m => m.default)
-      const { data = [] } = await octokit.request('GET /repos/{owner}/{repo}/commits', {
-        owner: 'vuetifyjs',
-        repo: '0',
-        per_page: 1,
-      })
-
-      if (data.length === 0) return
-
-      app.stats.commit = data[0] as typeof app.stats.commit
-      storage.set('latest', app.stats.commit)
-    } catch (error) {
-      logger.warn('Failed to fetch commit info', error)
-    }
-  })
+  onMounted(() => app.fetchCommit())
 </script>
 
 <template>
