+++ apps/docs/src/stores/app.ts
@@ -1,6 +1,12 @@
 import navData from 'virtual:nav'
 
+// Framework
+import { createStorage, useLogger } from '@vuetify/v0'
+
+import { CACHE_TTL } from '@/constants/cache'
+
 // Utilities
+import { flatten } from '@/utilities/nav'
 import { defineStore } from 'pinia'
 
 // Types
@@ -19,19 +25,7 @@ interface Commit {
   }
 }
 
-function flattenRoutes (nav: NavItem): string[] {
-  const routes: string[] = []
-
-  if ('to' in nav && nav.to) {
-    routes.push(nav.to)
-  }
-
-  if ('children' in nav && nav.children) {
-    routes.push(...nav.children.flatMap(child => flattenRoutes(child)))
-  }
-
-  return routes
-}
+const storage = createStorage({ prefix: 'v0-commit:', ttl: CACHE_TTL })
 
 export const useAppStore = defineStore('app', {
   state: () => ({
@@ -48,10 +42,41 @@ export const useAppStore = defineStore('app', {
       for (const nav of state.nav) {
         if (!('children' in nav) && !('to' in nav)) continue
 
-        pages.push(...flattenRoutes(nav as NavItem))
+        pages.push(...flatten(nav as NavItem))
       }
 
       return pages
+    },
+  },
+  actions: {
+    /**
+     * Latest commit on master, cached for `CACHE_TTL`. Every caller shares the
+     * cache — the footer used to re-hit the API on each intersection.
+     */
+    async fetchCommit () {
+      if (this.stats.commit) return
+
+      const cached = storage.get<Commit | null>('latest', null)
+      if (cached.value) {
+        this.stats.commit = cached.value
+        return
+      }
+
+      try {
+        const octokit = await import('@/plugins/octokit').then(m => m.default)
+        const { data = [] } = await octokit.request('GET /repos/{owner}/{repo}/commits', {
+          owner: 'vuetifyjs',
+          repo: '0',
+          per_page: 1,
+        })
+
+        if (data.length === 0) return
+
+        this.stats.commit = data[0] as Commit
+        storage.set('latest', this.stats.commit)
+      } catch (error) {
+        useLogger().warn('Failed to fetch commit info', error)
+      }
     },
   },
 })
